+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 - Added support for the Librem Key
+- Added support for user provided extensions through lookup via the
+  `PATH` environment variable
 - Added support for configuration files
   - Added support for configuration files that can be used to set
     default values for some arguments